+++ lithops/serverless/backends/aws_lambda/config.py
@@ -37,7 +37,8 @@ AVAILABLE_PY_RUNTIMES = {
     '3.8': 'python3.8',
     '3.9': 'python3.9',
     '3.10': 'python3.10',
-    '3.11': 'python3.11'
+    '3.11': 'python3.11',
+    '3.12': 'python3.12'
 }
 
 USER_RUNTIME_PREFIX = 'lithops.user_runtimes'